added j-tockAuth and change from seesion storage  to local storage
